requirements.txt
# -f https://download.pytorch.org/whl/torch_stable.html
# torch==1.7.1+cpu
# nltk==3.6.2
# transformers==4.19.2
# beautifulsoup4==4.9.3
# rubrix==0.13.0

# apache-airflow==2.2.2
apache-airflow-providers-mysql==2.2.3
apache-airflow-providers-postgres==4.1.0
apache-airflow-providers-amazon==3.4.0

awswrangler==2.15.1
boto3==1.21.1
botocore==1.24.21

pika==1.2.0
kafka-python==2.0.2
gdown==4.4.0